requirements.txt
aiogram==3.2.0
flask==2.3.3
flask-httpauth==4.8.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
tzdata>=2023.3
psycopg2-binary>=2.9.9